backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy==1.24.3
opencv-python==4.8.1.78
mediapipe==0.10.3  # Versión específica conocida por ser estable
pillow==10.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1